+++ mullvad-daemon/src/device/mod.rs
@@ -1100,11 +1100,18 @@ impl AccountManager {
         Ok(self.fetch_device_config(old_config))
     }
 
-    fn expiry_call(&self) -> Result<impl Future<Output = Result<DateTime<Utc>, Error>>, Error> {
+    fn expiry_call(
+        &self,
+    ) -> Result<impl Future<Output = Result<chrono::DateTime<Utc>, Error>>, Error> {
         let old_config = self.data.device().ok_or(Error::NoDevice)?;
         let account_token = old_config.account_token.clone();
         let account_service = self.account_service.clone();
-        Ok(async move { account_service.check_expiry_2(account_token).await })
+        Ok(async move {
+            account_service
+                .get_data_2(account_token)
+                .await
+                .map(|data| data.expiry)
+        })
     }
 
     fn needs_validation(&mut self) -> bool {