+++ CHANGELOG.md
@@ -23,6 +23,8 @@ Line wrap the file at 100 chars.                                              Th
 
 ## [Unreleased]
 ### Added
+- Add account UUID to verbose 'mullvad account get -v' output.
+
 #### Android
 - Add support for all screen orientations.